Ported rq job function to its own module, in order to be able to easily lazy-import packages only in case they are needed (essentially only when a job has really started). That way the main worker process that forks processes that execute the simulation jobs will not grow its memory utilization.
